applied eclipse 3.2M5 source cleanUp
git-svn-id: https://svn.apache.org/repos/asf/incubator/wicket/trunk/wicket@460002 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/wicket/util/lang/Objects.java
@@ -853,13 +853,13 @@ public abstract class Objects
 				return new Integer((int)value);
 
 			case FLOAT :
-				if ((long)value == value)
+				if (value == value)
 				{
 					return new Float(value);
 				}
 			// else fall through:
 			case DOUBLE :
-				if ((long)value == value)
+				if (value == value)
 				{
 					return new Double(value);
 				}
